# update csv to parquet to use multiprocessing